requirements.txt
discord
python-dotenv
docker
yt_dlp
asyncio
PyNaCl
